@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 import warnings
 from io import BytesIO
 from shutil import copyfileobj
@@ -27,35 +25,38 @@ class SqliteConnector(BaseDBConnector):
     Create a dump at SQL layer like could make ``.dumps`` in sqlite3.
     Restore by evaluate the created SQL.
     """
+
     def _write_dump(self, fileobj):
         cursor = self.connection.cursor()
         cursor.execute(DUMP_TABLES)
-        for table_name, type, sql in cursor.fetchall():
-            if table_name.startswith('sqlite_') or table_name in self.exclude:
+        for table_name, _, sql in cursor.fetchall():
+            if table_name.startswith("sqlite_") or table_name in self.exclude:
                 continue
-            elif sql.startswith('CREATE TABLE'):
-                sql = sql.replace('CREATE TABLE', 'CREATE TABLE IF NOT EXISTS')
+            if sql.startswith("CREATE TABLE"):
+                sql = sql.replace("CREATE TABLE", "CREATE TABLE IF NOT EXISTS")
                 # Make SQL commands in 1 line
-                sql = sql.replace('\n    ', '')
-                sql = sql.replace('\n)', ')')
-                fileobj.write("{};\n".format(sql).encode('UTF-8'))
-            else:
-                fileobj.write("{};\n".format(sql))
+                sql = sql.replace("\n    ", "")
+                sql = sql.replace("\n)", ")")
+            fileobj.write(f"{sql};\n".encode())
+
             table_name_ident = table_name.replace('"', '""')
-            res = cursor.execute('PRAGMA table_info("{0}")'.format(table_name_ident))
+            res = cursor.execute(f'PRAGMA table_info("{table_name_ident}")')
             column_names = [str(table_info[1]) for table_info in res.fetchall()]
             q = """SELECT 'INSERT INTO "{0}" VALUES({1})' FROM "{0}";\n""".format(
                 table_name_ident,
-                ",".join("""'||quote("{0}")||'""".format(col.replace('"', '""'))
-                         for col in column_names))
+                ",".join(
+                    """'||quote("{}")||'""".format(col.replace('"', '""'))
+                    for col in column_names
+                ),
+            )
             query_res = cursor.execute(q)
             for row in query_res:
-                fileobj.write("{};\n".format(row[0]).encode('UTF-8'))
+                fileobj.write(f"{row[0]};\n".encode())
             schema_res = cursor.execute(DUMP_ETC)
-            for name, type, sql in schema_res.fetchall():
+            for name, _, sql in schema_res.fetchall():
                 if sql.startswith("CREATE INDEX"):
-                    sql = sql.replace('CREATE INDEX', 'CREATE INDEX IF NOT EXISTS')
-                fileobj.write('{};\n'.format(sql).encode('UTF-8'))
+                    sql = sql.replace("CREATE INDEX", "CREATE INDEX IF NOT EXISTS")
+                fileobj.write(f"{sql};\n".encode())
         cursor.close()
 
     def create_dump(self):
@@ -72,11 +73,9 @@ class SqliteConnector(BaseDBConnector):
         cursor = self.connection.cursor()
         for line in dump.readlines():
             try:
-                cursor.execute(line.decode('UTF-8'))
-            except OperationalError as err:
-                warnings.warn("Error in db restore: {}".format(err))
-            except IntegrityError as err:
-                warnings.warn("Error in db restore: {}".format(err))
+                cursor.execute(line.decode("UTF-8"))
+            except (OperationalError, IntegrityError) as err:
+                warnings.warn(f"Error in db restore: {err}")
 
 
 class SqliteCPConnector(BaseDBConnector):
@@ -84,15 +83,16 @@ class SqliteCPConnector(BaseDBConnector):
     Create a dump by copy the binary data file.
     Restore by simply copy to the good location.
     """
+
     def create_dump(self):
-        path = self.connection.settings_dict['NAME']
+        path = self.connection.settings_dict["NAME"]
         dump = BytesIO()
-        with open(path, 'rb') as db_file:
+        with open(path, "rb") as db_file:
             copyfileobj(db_file, dump)
         dump.seek(0)
         return dump
 
     def restore_dump(self, dump):
-        path = self.connection.settings_dict['NAME']
-        with open(path, 'wb') as db_file:
+        path = self.connection.settings_dict["NAME"]
+        with open(path, "wb") as db_file:
             copyfileobj(dump, db_file)
