@@ -1,4 +1,5 @@
 import logging
+
 import django
 from django.utils.log import AdminEmailHandler
 
@@ -8,17 +9,22 @@ class DbbackupAdminEmailHandler(AdminEmailHandler):
         # Monkey patch for old Django versions without send_mail method
         if django.VERSION < (1, 8):
             from . import utils
+
             django.core.mail.mail_admins = utils.mail_admins
-        super(DbbackupAdminEmailHandler, self).emit(record)
+        super().emit(record)
 
     def send_mail(self, subject, message, *args, **kwargs):
         from . import utils
-        utils.mail_admins(subject, message, *args, connection=self.connection(), **kwargs)
+
+        utils.mail_admins(
+            subject, message, *args, connection=self.connection(), **kwargs
+        )
 
 
 class MailEnabledFilter(logging.Filter):
     def filter(self, record):
         from .settings import SEND_EMAIL
+
         return SEND_EMAIL
 
 
