@@ -1,3 +1,4 @@
+import contextlib
 import os
 import subprocess
 
@@ -8,19 +9,33 @@ from django.utils import timezone
 
 from dbbackup.db.base import get_connector
 
-BASE_FILE = os.path.join(settings.BLOB_DIR, 'test.txt')
-ENCRYPTED_FILE = os.path.join(settings.BLOB_DIR, 'test.txt.gpg')
-COMPRESSED_FILE = os.path.join(settings.BLOB_DIR, 'test.txt.gz')
-TARED_FILE = os.path.join(settings.BLOB_DIR, 'test.txt.tar')
-ENCRYPTED_COMPRESSED_FILE = os.path.join(settings.BLOB_DIR, 'test.txt.gz.gpg')
-TEST_DATABASE = {'ENGINE': 'django.db.backends.sqlite3', 'NAME': '/tmp/foo.db', 'USER': 'foo', 'PASSWORD': 'bar', 'HOST': 'foo', 'PORT': 122}
-TEST_MONGODB = {'ENGINE': 'django_mongodb_engine', 'NAME': 'mongo_test', 'USER': 'foo', 'PASSWORD': 'bar', 'HOST': 'foo', 'PORT': 122}
-TEST_DATABASE = settings.DATABASES['default']
+BASE_FILE = os.path.join(settings.BLOB_DIR, "test.txt")
+ENCRYPTED_FILE = os.path.join(settings.BLOB_DIR, "test.txt.gpg")
+COMPRESSED_FILE = os.path.join(settings.BLOB_DIR, "test.txt.gz")
+TARED_FILE = os.path.join(settings.BLOB_DIR, "test.txt.tar")
+ENCRYPTED_COMPRESSED_FILE = os.path.join(settings.BLOB_DIR, "test.txt.gz.gpg")
+TEST_DATABASE = {
+    "ENGINE": "django.db.backends.sqlite3",
+    "NAME": "/tmp/foo.db",
+    "USER": "foo",
+    "PASSWORD": "bar",
+    "HOST": "foo",
+    "PORT": 122,
+}
+TEST_MONGODB = {
+    "ENGINE": "django_mongodb_engine",
+    "NAME": "mongo_test",
+    "USER": "foo",
+    "PASSWORD": "bar",
+    "HOST": "foo",
+    "PORT": 122,
+}
+TEST_DATABASE = settings.DATABASES["default"]
 
-GPG_PRIVATE_PATH = os.path.join(settings.BLOB_DIR, 'gpg/secring.gpg')
-GPG_PUBLIC_PATH = os.path.join(settings.BLOB_DIR, 'gpg/pubring.gpg')
-GPG_FINGERPRINT = '7438 8D4E 02AF C011 4E2F  1E79 F7D1 BBF0 1F63 FDE9'
-DEV_NULL = open(os.devnull, 'w')
+GPG_PRIVATE_PATH = os.path.join(settings.BLOB_DIR, "gpg/secring.gpg")
+GPG_PUBLIC_PATH = os.path.join(settings.BLOB_DIR, "gpg/pubring.gpg")
+GPG_FINGERPRINT = "7438 8D4E 02AF C011 4E2F  1E79 F7D1 BBF0 1F63 FDE9"
+DEV_NULL = open(os.devnull, "w")
 
 
 class handled_files(dict):
@@ -29,21 +44,24 @@ class handled_files(dict):
     You should use the constant instance ``HANDLED_FILES`` and clean it
     before tests.
     """
+
     def __init__(self):
-        super(handled_files, self).__init__()
+        super().__init__()
         self.clean()
 
     def clean(self):
-        self['written_files'] = []
-        self['deleted_files'] = []
+        self["written_files"] = []
+        self["deleted_files"] = []
+
+
 HANDLED_FILES = handled_files()
 
 
 class FakeStorage(Storage):
-    name = 'FakeStorage'
+    name = "FakeStorage"
 
     def exists(self, name):
-        return name in HANDLED_FILES['written_files']
+        return name in HANDLED_FILES["written_files"]
 
     def get_available_name(self, name, max_length=None):
         return name[:max_length]
@@ -52,54 +70,47 @@ class FakeStorage(Storage):
         return name
 
     def listdir(self, path):
-        return ([], [f[0] for f in HANDLED_FILES['written_files']])
+        return ([], [f[0] for f in HANDLED_FILES["written_files"]])
 
     def accessed_time(self, name):
         return timezone.now()
+
     created_time = modified_time = accessed_time
 
-    def _open(self, name, mode='rb'):
-        file_ = [f[1] for f in HANDLED_FILES['written_files']
-                 if f[0] == name][0]
+    def _open(self, name, mode="rb"):
+        file_ = [f[1] for f in HANDLED_FILES["written_files"] if f[0] == name][0]
         file_.seek(0)
         return file_
 
     def _save(self, name, content):
-        HANDLED_FILES['written_files'].append((name, File(content)))
+        HANDLED_FILES["written_files"].append((name, File(content)))
         return name
 
     def delete(self, name):
-        HANDLED_FILES['deleted_files'].append(name)
+        HANDLED_FILES["deleted_files"].append(name)
 
 
 def clean_gpg_keys():
-    try:
-        cmd = ("gpg --batch --yes --delete-key '%s'" % GPG_FINGERPRINT)
+    with contextlib.suppress(Exception):
+        cmd = "gpg --batch --yes --delete-key '%s'" % GPG_FINGERPRINT
         subprocess.call(cmd, stdout=DEV_NULL, stderr=DEV_NULL)
-    except:
-        pass
-    try:
-        cmd = ("gpg --batch --yes --delete-secrect-key '%s'" % GPG_FINGERPRINT)
+    with contextlib.suppress(Exception):
+        cmd = "gpg --batch --yes --delete-secrect-key '%s'" % GPG_FINGERPRINT
         subprocess.call(cmd, stdout=DEV_NULL, stderr=DEV_NULL)
-    except:
-        pass
 
 
 def add_private_gpg():
-    cmd = ('gpg --import %s' % GPG_PRIVATE_PATH).split()
+    cmd = f"gpg --import {GPG_PRIVATE_PATH}".split()
     subprocess.call(cmd, stdout=DEV_NULL, stderr=DEV_NULL)
 
 
 def add_public_gpg():
-    cmd = ('gpg --import %s' % GPG_PUBLIC_PATH).split()
+    cmd = f"gpg --import {GPG_PUBLIC_PATH}".split()
     subprocess.call(cmd, stdout=DEV_NULL, stderr=DEV_NULL)
 
 
-
-
-
 def callable_for_filename_template(datetime, **kwargs):
-    return '%s_foo' % datetime
+    return f"{datetime}_foo"
 
 
 def get_dump(database=TEST_DATABASE):
