@@ -1,13 +1,14 @@
 """
 Tests for base command class.
 """
+
 import logging
 import os
 from io import BytesIO
+from unittest.mock import patch
 
 from django.core.files import File
 from django.test import TestCase
-from mock import patch
 
 from dbbackup.management.commands._base import BaseDbBackupCommand
 from dbbackup.storage import get_storage
@@ -51,25 +52,25 @@ class BaseDbBackupCommandMethodsTest(TestCase):
         self.command.storage = get_storage()
 
     def test_read_from_storage(self):
-        HANDLED_FILES['written_files'].append(['foo', File(BytesIO(b'bar'))])
-        file_ = self.command.read_from_storage('foo')
-        self.assertEqual(file_.read(), b'bar')
+        HANDLED_FILES["written_files"].append(["foo", File(BytesIO(b"bar"))])
+        file_ = self.command.read_from_storage("foo")
+        self.assertEqual(file_.read(), b"bar")
 
     def test_write_to_storage(self):
-        self.command.write_to_storage(BytesIO(b'foo'), 'bar')
-        self.assertEqual(HANDLED_FILES['written_files'][0][0], 'bar')
+        self.command.write_to_storage(BytesIO(b"foo"), "bar")
+        self.assertEqual(HANDLED_FILES["written_files"][0][0], "bar")
 
     def test_read_local_file(self):
         # setUp
-        self.command.path = '/tmp/foo.bak'
-        open(self.command.path, 'w').close()
+        self.command.path = "/tmp/foo.bak"
+        open(self.command.path, "w").close()
         # Test
-        output_file = self.command.read_local_file(self.command.path)
+        self.command.read_local_file(self.command.path)
         # tearDown
         os.remove(self.command.path)
 
     def test_write_local_file(self):
-        fd, path = File(BytesIO(b"foo")), '/tmp/foo.bak'
+        fd, path = File(BytesIO(b"foo")), "/tmp/foo.bak"
         self.command.write_local_file(fd, path)
         self.assertTrue(os.path.exists(path))
         # tearDown
@@ -77,22 +78,22 @@ class BaseDbBackupCommandMethodsTest(TestCase):
 
     def test_ask_confirmation(self):
         # Yes
-        with patch('dbbackup.management.commands._base.input', return_value='y'):
+        with patch("dbbackup.management.commands._base.input", return_value="y"):
             self.command._ask_confirmation()
-        with patch('dbbackup.management.commands._base.input', return_value='Y'):
+        with patch("dbbackup.management.commands._base.input", return_value="Y"):
             self.command._ask_confirmation()
-        with patch('dbbackup.management.commands._base.input', return_value=''):
+        with patch("dbbackup.management.commands._base.input", return_value=""):
             self.command._ask_confirmation()
-        with patch('dbbackup.management.commands._base.input', return_value='foo'):
+        with patch("dbbackup.management.commands._base.input", return_value="foo"):
             self.command._ask_confirmation()
         # No
-        with patch('dbbackup.management.commands._base.input', return_value='n'):
+        with patch("dbbackup.management.commands._base.input", return_value="n"):
             with self.assertRaises(SystemExit):
                 self.command._ask_confirmation()
-        with patch('dbbackup.management.commands._base.input', return_value='N'):
+        with patch("dbbackup.management.commands._base.input", return_value="N"):
             with self.assertRaises(SystemExit):
                 self.command._ask_confirmation()
-        with patch('dbbackup.management.commands._base.input', return_value='No'):
+        with patch("dbbackup.management.commands._base.input", return_value="No"):
             with self.assertRaises(SystemExit):
                 self.command._ask_confirmation()
 
@@ -104,52 +105,59 @@ class BaseDbBackupCommandCleanupOldBackupsTest(TestCase):
         self.command.stdout = DEV_NULL
         self.command.encrypt = False
         self.command.compress = False
-        self.command.servername = 'foo-server'
+        self.command.servername = "foo-server"
         self.command.storage = get_storage()
-        HANDLED_FILES['written_files'] = [(f, None) for f in [
-            'fooserver-2015-02-06-042810.tar',
-            'fooserver-2015-02-07-042810.tar',
-            'fooserver-2015-02-08-042810.tar',
-            'foodb-fooserver-2015-02-06-042810.dump',
-            'foodb-fooserver-2015-02-07-042810.dump',
-            'foodb-fooserver-2015-02-08-042810.dump',
-            'bardb-fooserver-2015-02-06-042810.dump',
-            'bardb-fooserver-2015-02-07-042810.dump',
-            'bardb-fooserver-2015-02-08-042810.dump',
-            'hamdb-hamserver-2015-02-06-042810.dump',
-            'hamdb-hamserver-2015-02-07-042810.dump',
-            'hamdb-hamserver-2015-02-08-042810.dump',
-        ]]
+        HANDLED_FILES["written_files"] = [
+            (f, None)
+            for f in [
+                "fooserver-2015-02-06-042810.tar",
+                "fooserver-2015-02-07-042810.tar",
+                "fooserver-2015-02-08-042810.tar",
+                "foodb-fooserver-2015-02-06-042810.dump",
+                "foodb-fooserver-2015-02-07-042810.dump",
+                "foodb-fooserver-2015-02-08-042810.dump",
+                "bardb-fooserver-2015-02-06-042810.dump",
+                "bardb-fooserver-2015-02-07-042810.dump",
+                "bardb-fooserver-2015-02-08-042810.dump",
+                "hamdb-hamserver-2015-02-06-042810.dump",
+                "hamdb-hamserver-2015-02-07-042810.dump",
+                "hamdb-hamserver-2015-02-08-042810.dump",
+            ]
+        ]
 
-    @patch('dbbackup.settings.CLEANUP_KEEP', 1)
+    @patch("dbbackup.settings.CLEANUP_KEEP", 1)
     def test_clean_db(self):
-        self.command.content_type = 'db'
-        self.command.database = 'foodb'
-        self.command._cleanup_old_backups(database='foodb')
-        self.assertEqual(2, len(HANDLED_FILES['deleted_files']))
-        self.assertNotIn('foodb-fooserver-2015-02-08-042810.dump',
-                         HANDLED_FILES['deleted_files'])
+        self.command.content_type = "db"
+        self.command.database = "foodb"
+        self.command._cleanup_old_backups(database="foodb")
+        self.assertEqual(2, len(HANDLED_FILES["deleted_files"]))
+        self.assertNotIn(
+            "foodb-fooserver-2015-02-08-042810.dump", HANDLED_FILES["deleted_files"]
+        )
 
-    @patch('dbbackup.settings.CLEANUP_KEEP', 1)
+    @patch("dbbackup.settings.CLEANUP_KEEP", 1)
     def test_clean_other_db(self):
-        self.command.content_type = 'db'
-        self.command._cleanup_old_backups(database='bardb')
-        self.assertEqual(2, len(HANDLED_FILES['deleted_files']))
-        self.assertNotIn('bardb-fooserver-2015-02-08-042810.dump',
-                         HANDLED_FILES['deleted_files'])
+        self.command.content_type = "db"
+        self.command._cleanup_old_backups(database="bardb")
+        self.assertEqual(2, len(HANDLED_FILES["deleted_files"]))
+        self.assertNotIn(
+            "bardb-fooserver-2015-02-08-042810.dump", HANDLED_FILES["deleted_files"]
+        )
 
-    @patch('dbbackup.settings.CLEANUP_KEEP', 1)
+    @patch("dbbackup.settings.CLEANUP_KEEP", 1)
     def test_clean_other_server_db(self):
-        self.command.content_type = 'db'
-        self.command._cleanup_old_backups(database='bardb')
-        self.assertEqual(2, len(HANDLED_FILES['deleted_files']))
-        self.assertNotIn('bardb-fooserver-2015-02-08-042810.dump',
-                         HANDLED_FILES['deleted_files'])
+        self.command.content_type = "db"
+        self.command._cleanup_old_backups(database="bardb")
+        self.assertEqual(2, len(HANDLED_FILES["deleted_files"]))
+        self.assertNotIn(
+            "bardb-fooserver-2015-02-08-042810.dump", HANDLED_FILES["deleted_files"]
+        )
 
-    @patch('dbbackup.settings.CLEANUP_KEEP_MEDIA', 1)
+    @patch("dbbackup.settings.CLEANUP_KEEP_MEDIA", 1)
     def test_clean_media(self):
-        self.command.content_type = 'media'
+        self.command.content_type = "media"
         self.command._cleanup_old_backups()
-        self.assertEqual(2, len(HANDLED_FILES['deleted_files']))
-        self.assertNotIn('foo-server-2015-02-08-042810.tar',
-                         HANDLED_FILES['deleted_files'])
+        self.assertEqual(2, len(HANDLED_FILES["deleted_files"]))
+        self.assertNotIn(
+            "foo-server-2015-02-08-042810.tar", HANDLED_FILES["deleted_files"]
+        )
