@@ -1,9 +1,8 @@
-from __future__ import unicode_literals
-
 from io import BytesIO
+from unittest.mock import mock_open, patch
 
+from django.db import connection
 from django.test import TestCase
-from mock import mock_open, patch
 
 from dbbackup.db.sqlite import SqliteConnector, SqliteCPConnector
 from dbbackup.tests.testapp.models import CharModel
@@ -16,7 +15,7 @@ class SqliteConnectorTest(TestCase):
         connector._write_dump(dump_file)
         dump_file.seek(0)
         for line in dump_file:
-            self.assertTrue(line.strip().endswith(b';'))
+            self.assertTrue(line.strip().endswith(b";"))
 
     def test_create_dump(self):
         connector = SqliteConnector()
@@ -24,7 +23,7 @@ class SqliteConnectorTest(TestCase):
         self.assertTrue(dump.read())
 
     def test_create_dump_with_unicode(self):
-        CharModel.objects.create(field='\xe9')
+        CharModel.objects.create(field="\xe9")
         connector = SqliteConnector()
         dump = connector.create_dump()
         self.assertTrue(dump.read())
@@ -34,15 +33,23 @@ class SqliteConnectorTest(TestCase):
         dump = connector.create_dump()
         connector.restore_dump(dump)
 
+    def test_create_dump_with_virtual_tables(self):
+        with connection.cursor() as c:
+            c.execute("CREATE VIRTUAL TABLE lookup USING fts5(field)")
 
-@patch('dbbackup.db.sqlite.open', mock_open(read_data=b'foo'), create=True)
+        connector = SqliteConnector()
+        dump = connector.create_dump()
+        self.assertTrue(dump.read())
+
+
+@patch("dbbackup.db.sqlite.open", mock_open(read_data=b"foo"), create=True)
 class SqliteCPConnectorTest(TestCase):
     def test_create_dump(self):
         connector = SqliteCPConnector()
         dump = connector.create_dump()
         dump_content = dump.read()
         self.assertTrue(dump_content)
-        self.assertEqual(dump_content, b'foo')
+        self.assertEqual(dump_content, b"foo")
 
     def test_restore_dump(self):
         connector = SqliteCPConnector()
